requirements.txt
fastapi==0.110.3             # FastAPI 웹 프레임워크, 라우팅/요청처리/응답 지원
uvicorn==0.29.0              # ASGI 서버 (FastAPI 앱 실행)
python-multipart==0.0.9      # 파일 업로드(FormData) 지원
pydantic==2.6.4              # 요청/응답 스키마 정의 및 검증
python-dotenv==1.0.1         # .env 파일 로드해 환경변수 설정

pypdf==4.2.0                 # PDF 텍스트 추출 (PdfReader 사용)
docx2txt>=0.8                # DOCX 텍스트 추출 (로봇 스펙 문서 업로드용)
pdf2image>=1.17              # 스캔 PDF → 이미지 변환 (OCR용, poppler 필요)
pytesseract>=0.3.10          # OCR (tesseract-ocr 설치 필요)

langchain>=0.1.16             # LangChain 메타 패키지
langchain-core>=0.1.0        # LCEL, 프롬프트
langchain-openai>=0.1.7      # OpenAI / GPT-4o Vision 연동
langchain-community>=0.0.20  # FAISS 벡터스토어
langchain-text-splitters>=0.0.1  # 텍스트 분할

faiss-cpu>=1.7.4             # FAISS 기반 지식베이스 (선택 설치 가능)

pillow==10.4.0               # 히트맵/오버레이 이미지 생성
PyJWT>=2.8.0                 # JWT 로컬 검증 (deps.py)